requirements.txt
Flask==2.0.2
line-bot-sdk==1.17.0
